Record image rendering FPS at set_next_image_all() and set_prev_image_all()
To avoid possible double-counting in dual-pane mode.
